Extend `.gitignore` to include IntelliJ files
This actually uses http://gitignore.io to generate an ignore file
for both Java and IntelliJ. I chose to *not* ignore the IML and a
few of the XML files as per the recommendations from IntelliJ
(https://intellij-support.jetbrains.com/hc/en-us/articles/206544839)
and this issue post
(https://github.com/joeblau/gitignore.io/issues/186#issuecomment-215987721)
that points out that without these IML/XML files you have to reconstruct
the modules which is kind of a pain.

I've also gone ahead and added the unignored IntelliJ files to the
repository.