Move `EventList` into its own file
Used the IntelliJ refactoring tools to move `EventList` into its own
files. This required creating a getter for the array `finishTime` which
is then used in the `MachineShopSimulator` to actually *set* values in
that array. This indicates that this array is not properly encapsulated
and we should come back and clean that up.