Extract all the inner classes from `MachineShopSimulator` (#36)
* Encapsulated the fields in the `Task` class

Used the IntelliJ encapsulation refactoring to properly encapsulate
the fields in the `Task` class so that other classes are only accessing
them through the accessors instead of taking advantage of the smelly
use of inner classes.

* Move `Task` into its own file

Used the IntelliJ move refactoring to move this inner class into its
own file.

* Move `Job` into its own file

Used the IntelliJ field encapsulation refactoring to encapsulate all the
fields in `Job`. This included the need for a setter for `arrivalTime`
since it is updated by the `MachineShopSimulator`. Then used the
move refactoring to move this inner class into its own file.

* Move `Machine` into its own file

Moved `Machine` into its own file. Turns out that all the fields there
were package visible so I encapsulated them using the IntelliJ
encapsulation refactoring tool.

* Move `EventList` into its own file

Used the IntelliJ refactoring tools to move `EventList` into its own
files. This required creating a getter for the array `finishTime` which
is then used in the `MachineShopSimulator` to actually *set* values in
that array. This indicates that this array is not properly encapsulated
and we should come back and clean that up.