Create input and output objects and begin parsing into those
Currently input is all prompt-driven reading from standard input
and output is scattered across a couple of methods. This refactoring
begins to construct input and output classes that can be used to
more easily test the simulator by allowing us to pass in a populated
input object and return an output object that we can check against the
expected results.

This currently populates the entire `SimulationSpecification` object
in `readData()`. It also does a lot of the setup there, which needs
to be moved to `startShop()`.

The use of 1-indexed arrays is incredibly annoying and has led to
several bugs that were, fortunately, caught by the tests. (Hopefully
there are no other similar errors that the tests didn't catch.)

We were then able to extract all the "setup" logic from `inputData()`
and move that to the `startShop()` method, which gets its necessary
input values from the `SimulationSpecification` input object it is
passed. This completely separates the reading and parsing code from
the simulation setup code.