Extract all printing into `SimulationResults`
This extracts all the printing out of the `MachineShopSimulator` and
into the new `SimulationResults` class. Now the simulator has no
direct connection to any of the I/O; all the reading happens in
`SpecificationReader` and all the printing happens in
`SimulationResults`. This will allow us to create tests that construct
`SimulationSpecification` objects as inputs to the simulator and look
at `SimulationResults` objects as outputs from the simulator.

We created a small, new `JobCompletionData` class to capture various
pieces of important information when a job finishes.