Move all the tests into a `tests` directory
I had the unit tests and the "acceptance" tests in two different
directories, which made it more difficult to treat all the tests as a
group in several ways. So I created a new `tests` directory and made
both of those previous directories sub-directories there.

This also required a minor change to `FileIoTests` to indicate the
new location of the acceptance test directory.