Add `toString()` and `size()` methods to `LinkedQueue`
When writing some properties for `LinkedQueue` I found it useful to
have a meaningful `toString()` and a `size()` method in `LinkedQueue`.
The `toString()` was useful so that I could print queues that failed
tests and see what kinds of queues led to problems. The `size()` was
useful in implementing a meaningful `magnitude()` method in my generator,
which turned out to be critical to getting shrinking to work.