Add a getter for `SimulationResults.totalWaitTimePerMachine`
We need to know the total wait times per machine for some of the
property-based testing, so I added a getter for that. Because that's
an array, I returned a *copy* of it to make sure that if someone
messed with it when they got it, that wouldn't make changes back in
the results object.