Add a `toString()` to `SimulationSpecification`
When property-based tests fail it would be helpful to print out the
input specification that led to the failure, so I added a `toString()`
method. The generated string can be fairly long, especially if there
are a lot of machines and jobs, but hopefully test case shrinking will
make that less of a concern.