Add `doShrink()` and `magnitude()` to `SimulationSpecificationGenerator`
Now that I have some non-trivial properties, I needed to add
`doShrink()` and `magnitude()` methods to the generator so that
`junit-quickcheck` will actually attempt to shrink the test inputs
and give us something more readable and easier to understand.

My `doShrink()` attempts to create two new specifications, one by
removing a random machine (& all references to it in the task lists),
and one by removing a random task and it's task list. If either of
these leads to an "impossible" state (no machines, no jobs, or no
tasks on a job) then we don't get that new specification.

My `magnitude()` is the sum of

   * the number of machines
   * the number of jobs
   * the total of the number of tasks across all the jobs

I suspect we could come up with something fancier here, but this
`doShrink()` and `magnitude()` do seem to get us down to one machine
and one job in most of the examples I tried, so I think that's pretty
good.